backend/requirements.txt
aiofiles==0.7.0
aioredis
boto3==1.28.63
botocore==1.31.63
fastapi==0.103.2
fastapi-redis-cache
httpx
itsdangerous
minio
openai==1.46.0
openpyxl==3.0.7
pandas
passlib
psycopg2-binary
PyMuPDF
PyPDF2==1.26.0
python-docx==0.8.11
python-dotenv==0.19.0
python-jose~=3.3.0
python-multipart
requests==2.26.0
sqlalchemy
uvicorn==0.23.2
uuid
httpx
requests
